+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: Alias field in plugin should be disabled unless site/category is populated
 
 1.7.0 (2022-05-31)
 ==================